[feat] : BottleItem UI 구현
